revert

+++ app.py
@@ -6,93 +6,103 @@ from werkzeug.exceptions import NotFound
 import os
 
 host = os.environ.get('MONGODB_URI', 'mongodb://localhost:27017/Playlister')
+client = MongoClient(host=f'{host}?retryWrites=false')
+db = client.get_default_database()
 
-app = Flask(__name__)  # Flask app
-client = MongoClient(host=f'{host}?retryWrites=false')  # Mongo client
-
-db = client.get_default_database()  # New db with name of Playlister
-playlists = db.playlists  # create playlists comment
+playlists = db.playlists
 comments = db.comments
+
+
+app = Flask(__name__)
 
 
 @app.route('/')
 def playlists_index():
     """Show all playlists."""
-    return render_template('playlists_index.html', playlists=playlists.find())  # Return an iterable of all playlists in db
+    playlist = playlists.find()
+    # This will display all playlists by looping through the database
+    return render_template('playlists_index.html', playlist=playlist)
 
 
 @app.route('/playlists', methods=['POST'])
-def paylists_submit():
+def playlists_submit():
     """Submit a new playlist."""
-    #playlist = request.form.to_dict()
-    playlist = { 
+    playlist = {
         'title': request.form.get('title'),
         'description': request.form.get('description'),
-        'rating': request.form.get('rating'),
         'videos': request.form.get('videos').split(),
-        'created_at': datetime.datetime.today()
+        'created_at': datetime.now()
     }
+    print(playlist)
     playlist_id = playlists.insert_one(playlist).inserted_id
     return redirect(url_for('playlists_show', playlist_id=playlist_id))
 
 
 @app.route('/playlists/new')
 def playlists_new():
-    """Create a new playlist"""
+    """Create a new playlist."""
+    # This route will create a new playlist and display the playist_form.html
     return render_template('playlists_new.html', playlist={}, title='New Playlist')
+
+
+@app.route('/playlists/<id>/edit')
+def playlists_edit(playlist_id):
+    """Show the edit form for a playlist."""
+    # This will edit a specific playlist by its id
+    playlist = playlists.find_one({'_id': ObjectId(playlist_id)})
+    video_links = '\n'.join(playlist.get('videos'))
+    return render_template('playlists_edit.html', playlist=playlist)
 
 
 @app.route('/playlists/<playlist_id>')
 def playlists_show(playlist_id):
     """Show a single playlist."""
-    playlist_id = ObjectId(playlist_id)
-    playlist = playlists.find_one({'_id': playlist_id})
-    playlist_comments = comments.find({'playlist_id': playlist_id})
+    playlist = playlists.find_one({'_id': ObjectId(playlist_id)})
+    playlist_comments = comments.find({'playlist_id': ObjectId(playlist_id)})
     return render_template('playlists_show.html', playlist=playlist, comments=playlist_comments)
 
 
 @app.route('/playlists/<playlist_id>', methods=['POST'])
+# This will call on either to display or edit a playlist based on what the user submits
 def playlists_update(playlist_id):
     """Submit an edited playlist."""
-    updated_playlist = {
-        'title': request.form.get('title'),
-        'description': request.form.get('description'),
-        'rating': request.form.get('rating'),
-        'videos': request.form.get('videos').split(),
-        'created_at': datetime.datetime.today()
-    }
-    playlists.update_one({'_id': ObjectId(playlist_id)},
-                         {'$set': updated_playlist})
-    return redirect(url_for('playlists_show', playlist_id=playlist_id))
-
-
-@app.route('/playlists/<playlist_id>/edit')
-def playlists_edit(playlist_id):
-    # TODO: Have rating value auto select instead of reseting back to 1 in edit form.
-    """Show the edit form for a playlist."""
-    playlist = playlists.find_one({'_id': ObjectId(playlist_id)})
-    video_links = '\n'.join(playlist.get('videos'))  # Huh?
-    return render_template('playlists_edit.html', playlist=playlist, title='Edit Playlist')
+    if request.form.get('_method') == 'PUT':
+        updated_playlist = {
+            'title': request.form.get('title'),
+            'description': request.form.get('description'),
+            'videos': request.form.get('videos').split()
+        }
+        playlists.update_one(
+            {'_id': ObjectId(playlist_id)},
+            {'$set': updated_playlist})
+        return redirect(url_for('playlists_show', playlist_id=playlist_id))
+    else:
+        raise NotFound()
 
 
 @app.route('/playlists/<playlist_id>/delete', methods=['POST'])
 def playlists_delete(playlist_id):
+    # This will delete a playlist by using an id as a parameter
     """Delete one playlist."""
-    playlists.delete_one({'_id': ObjectId(playlist_id)})
-    return redirect(url_for('playlists_index'))
+    if request.form.get('_method') == 'DELETE':
+        playlists.delete_one({'_id': ObjectId(playlist_id)})
+        return redirect(url_for('playlists_index'))
+    else:
+        raise NotFound()
 
 
 @app.route('/playlists/comments', methods=['POST'])
 def comments_new():
     """Submit a new comment."""
-    playlist_id = ObjectId(request.form.get('playlist_id'))
     comment = {
         'title': request.form.get('title'),
         'content': request.form.get('content'),
-        'playlist_id': playlist_id
+        'playlist_id': ObjectId(request.form.get('playlist_id'))
     }
+    print(comment)
     comment_id = comments.insert_one(comment).inserted_id
-    return redirect(url_for('playlists_show', playlist_id=playlist_id))
+    return redirect(url_for('playlists_show', comment_id=comment_id))
+
 
 @app.route('/playlists/comments/<comment_id>', methods=['POST'])
 def comments_delete(comment_id):
